icer.com.pl/zaplecze/requirements.txt
absl-py==1.4.0
aiohttp==3.9.4
aiosignal==1.3.1
APScheduler==3.10.4
astunparse==1.6.3
attrs==23.2.0
barcode==1.0.2
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.51.0
frozenlist==1.4.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
Levenshtein==0.21.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.4
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
netifaces==0.10.6
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.2
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-engineio==4.8.2
python-Levenshtein==0.21.1
python-socketio==5.11.0
pytz==2024.1
pyzbar==0.1.9
qrcode==7.4.2
rapidfuzz==3.2.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2024.1
tzlocal==5.2
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.16
Werkzeug==2.3.7
wikipedia==1.4.0
Wikipedia-API==0.6.0
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.4
